+++ client.py
@@ -1,0 +1,6 @@
+import discord
+                
+me_st = discord.Game("battles! ⚔")
+intents = discord.Intents.default()
+intents.members = True
+client = discord.Client(intents=intents, activity=me_st)
